+++ packages/core/src/components/account/orders/MyAccountOrderDetails/MyAccountOrderDetails.tsx
@@ -6,7 +6,9 @@ import { MyAccountDeliveryOptionAccordion } from './MyAccountDeliveryOptionAccor
 import MyAccountOrderActions from './MyAccountOrderActions'
 import MyAccountOrderedByCard from './MyAccountOrderedByCard'
 import MyAccountPaymentCard from './MyAccountPaymentCard'
+import MyAccountReorderButton from './MyAccountReorderButton'
 import MyAccountSummaryCard from './MyAccountSummaryCard'
+import MyAccountBudgetsCard from './MyAccountBudgetsCard'
 
 import type {
   ServerOrderDetailsQueryQuery,
@@ -55,11 +57,15 @@ export default function MyAccountOrderDetails({
           </div>
         </div>
 
-        <MyAccountOrderActions
-          allowCancellation={order.allowCancellation}
-          orderId={order.orderId}
-          customerEmail={order.clientProfileData?.email}
-        />
+        <div data-fs-order-details-header-actions>
+          <MyAccountReorderButton order={order} />
+          <MyAccountOrderActions
+            allowCancellation={order.allowCancellation}
+            orderId={order.orderId}
+            customerEmail={order.clientProfileData?.email}
+            order={order}
+          />
+        </div>
       </header>
 
       <main data-fs-order-details-content>
@@ -116,6 +122,13 @@ export default function MyAccountOrderDetails({
 
         {moreInformationCustomFields?.length > 0 && (
           <MyAccountMoreInformationCard fields={moreInformationCustomFields} />
+        )}
+
+        {order.budgetData && (
+          <MyAccountBudgetsCard
+            budgetData={order.budgetData}
+            currencyCode={order.storePreferencesData.currencyCode}
+          />
         )}
       </main>
     </div>