+++ packages/core/src/sdk/error/ReorderError/ReorderError.ts
@@ -1,0 +1,6 @@
+export default class ReorderError extends Error {
+  constructor(message: string) {
+    super(message)
+    this.name = 'ReorderError'
+  }
+}